pyproject.toml
[tool.poetry]
name = "hackday-tour-embeddings"
version = "0.1.0"
description = ""
authors = ["Duy Pham <duy.pham@getyourguide.com>"]
readme = "README.md"


[[tool.poetry.source]]
name = "codeartifact"
url = "https://getyourguide-130607246975.d.codeartifact.eu-central-1.amazonaws.com/pypi/private/simple/"
priority = "explicit"


[tool.poetry.dependencies]
python = "^3.10"
pyspark = "^3.5.1"
pandas = "~1.5.3"
fire = "^0.6.0"
tensorboard = "^2.15.1"
pytorch-ignite = "^0.5.1"
numpy = "1.23.*"
torch = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"
pylint = "^3.2.2"
databricks-rocket = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
